requirements.txt
# Core
Django==5.2.4
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
sqlparse==0.5.3

# Async / Channels
asgiref==3.9.1
channels==4.3.1
channels_redis==4.3.0
daphne==4.2.1
Twisted==25.5.0
autobahn==24.4.2
txaio==25.6.1
zope.interface==8.0
service-identity==24.2.0
Automat==25.4.16
constantly==23.10.4
incremental==24.7.2
hyperlink==21.0.0

# Celery + Redis + Scheduling
celery[redis]==5.5.3
django-celery-beat==2.8.1
redis==5.2.1
kombu==5.5.2
vine==5.1.0
billiard==4.2.1

# Database
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tzlocal==5.3.1

# Utils
python-decouple==3.8
phonenumbers==9.0.9
django-phonenumber-field==8.1.0
dateparser==1.2.2
regex==2025.7.34
inflection==0.5.1
packaging==25.0

# Serialization / Validation
pydantic==2.11.7
pydantic_core==2.33.2
typing_extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0

# HTTP Clients
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
anyio==4.9.0
sniffio==1.3.1
certifi==2025.7.14
charset-normalizer==3.4.2
urllib3==2.5.0
idna==3.10

# OpenAI + AI Tools
openai==1.96.1
jiter==0.10.0
numpy==2.3.2
tqdm==4.67.1

# Email / SMS
sendgrid==6.12.4
python-http-client==3.3.7
twilio==9.6.5
ecdsa==0.19.1
pyasn1==0.6.1
pyasn1_modules==0.4.2

# Geo / Location
geopy==2.4.1
geographiclib==2.0
h3==4.3.0
timezonefinder==7.0.2

# Serialization / Formats
PyYAML==6.0.2
uritemplate==4.2.0
MarkupSafe==3.0.2

# Async Utils
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
frozenlist==1.7.0
multidict==6.6.3
attrs==25.3.0
yarl==1.20.1
propcache==0.3.2
aiohappyeyeballs==2.6.1

# Security / Crypto
cryptography==45.0.5
pyOpenSSL==25.1.0
cffi==1.17.1
pycparser==2.22

# Media
pillow==11.3.0

# Deployment
gunicorn==23.0.0
Werkzeug==3.1.3
setuptools==80.9.0

openpyxl==3.1.5